Major refactoring underway

+++ src/main/java/org/blackcat/trunk/http/requests/handlers/impl/DeleteResourceRequestHandlerImpl.java
@@ -1,5 +1,6 @@
 package org.blackcat.trunk.http.requests.handlers.impl;
 
+import io.vertx.core.json.JsonObject;
 import io.vertx.core.logging.Logger;
 import io.vertx.core.logging.LoggerFactory;
 import io.vertx.ext.web.RoutingContext;
@@ -18,31 +19,32 @@ final public class DeleteResourceRequestHandlerImpl extends BaseUserRequestHandl
     public void handle(RoutingContext ctx) {
         super.handle(ctx);
 
-        Path protectedPath = protectedPath(ctx);
-        Path resolvedPath = storage.getRoot().resolve(protectedPath);
-        logger.trace("DELETE {} -> {}", protectedPath, resolvedPath);
+        if (ctx.get("requestType").equals(RequestType.HTML))
+            jsonResponseBuilder.badRequest(ctx);
+        else {
+            Path protectedPath = protectedPath(ctx);
+            Path resolvedPath = storage.getRoot().resolve(protectedPath);
+            logger.trace("DELETE {} -> {}", protectedPath, resolvedPath);
 
-        storage.delete(resolvedPath, resource -> {
-            if (resource instanceof ErrorResource) {
-                ErrorResource errorResource = (ErrorResource) resource;
-                if (errorResource.isNotFound()) {
-                    responseBuilder.notFound(ctx);
-                    return;
+            storage.delete(resolvedPath, resource -> {
+                if (resource instanceof ErrorResource) {
+                    ErrorResource errorResource = (ErrorResource) resource;
+                    if (errorResource.isNotFound()) {
+                        jsonResponseBuilder.notFound(ctx);
+                        return;
+                    } else if (errorResource.isInvalid()) {
+                        jsonResponseBuilder.conflict(ctx);
+                        return;
+                    } else if (errorResource.isUnit()) {
+                        logger.debug("Ok: {}", ctx.request().uri());
+                        jsonResponseBuilder.success(ctx, new JsonObject());
+                        return;
+                    }
                 }
 
-                else if (errorResource.isInvalid()) {
-                    responseBuilder.conflict(ctx, errorResource.getMessage());
-                    return;
-                }
-
-                else if (errorResource.isUnit()) {
-                    responseBuilder.ok(ctx);
-                    return;
-                }
-            }
-
-            logger.error("Unexpected error resource type");
-            responseBuilder.internalServerError(ctx);
-        });
+                logger.error("Unexpected error resource type");
+                ctx.fail(new BaseUserRequestException("Unexpected error resource type"));
+            });
+        }
     }
 }
